# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handler.ComputerReq:
    properties:
      cpu:
        type: string
      gpu:
        type: string
      hdd:
        type: string
      ip:
        type: string
      manufacturer:
        type: string
      os:
        type: string
      ram:
        type: string
    type: object
  handler.UserReq:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
    type: object
  responder.Response:
    properties:
      code:
        type: integer
      contentType:
        type: string
      payload: {}
    type: object
...
host: 192.168.49.2:31532
info:
  contact: {}
  description: Practice API
  title: Practice
  version: "1.0"
paths:
...
  /computer/rabbit:
    post:
      consumes:
      - application/json
      description: Creates a computer instance via RabbitMQ
      parameters:
      - description: Computer object
        in: body
        name: computer
        required: true
        schema:
          $ref: '#/definitions/handler.ComputerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: Computer creation through RabbitMQ
      tags:
      - RabbitMQ
  /computer/rabbit/{id}:
    delete:
      description: Deletes a computer instance via RabbitMQ
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: Computer deletion through RabbitMQ
      tags:
      - RabbitMQ
    put:
      description: Updates a computer instance via RabbitMQ
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: string
      - description: Computer object
        in: body
        name: computer
        required: true
        schema:
          $ref: '#/definitions/handler.ComputerReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: Computer update through RabbitMQ
      tags:
      - RabbitMQ
...
  /user/rabbit:
    post:
      consumes:
      - application/json
      description: Adds a new user instance via RabbitMQ
      parameters:
      - description: User object
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/handler.UserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: User creation through RabbitMQ
      tags:
      - RabbitMQ
  /user/rabbit/{id}:
    delete:
      description: Deletes a user instance via RabbitMQ
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: User deletion through RabbitMQ
      tags:
      - RabbitMQ
    put:
      description: Updates a user instance via RabbitMQ
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/handler.UserReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responder.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responder.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responder.Response'
      summary: User update through RabbitMQ
      tags:
      - RabbitMQ
swagger: "2.0"